% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_window.R
\name{acf_window}
\alias{acf_window}
\title{Auto Correlation with a moving window}
\usage{
acf_window(df = df_with_windows, multipeak_period = TRUE)
}
\arguments{
\item{df}{A data.frame with 2 columns. Column 1 must contain the windows to iterate over.
Column 2 must supply the values. This parameter is optional if window_vector and values are supplied.
df must not have gaps in the dates, acf asumes data is evenly spaced.}

\item{multipeak_period}{TRUE (default) use all positive peaks to find the period.
FALSE use peak_of_interest to find period.}

\item{peak_of_interest}{Positive peak on which we want to base the period calculation.
If not peak is supplied it will use the peak next to the middle peak.}

\item{window_vector}{A vector containing the windows to iterate over and to label the group
to which each value belongs. Usually this will be the output from make_time_windows.}

\item{values}{The data to which we want to find the period.}
}
\value{
A data.frame with the autocorrelation results for each window which include: period, peaks,
power, lags for the peaks.
}
\description{
Iterated autocorrelation over windowed (grouped) timeseries data.
}
\examples{
autocorrelations_multipeak <- acf_window(df = df_with_windows, multipeak_period = TRUE)

}
\seealso{
[stats::acf()] which this functions uses to run the autocorrelation.
}
