% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_timeseries.grangertest.R
\name{analyze_timeseries.grangertest}
\alias{analyze_timeseries.grangertest}
\title{Apply the Granger Causality test on both directions}
\usage{
analyze_timeseries.grangertest(value, cos, order = 2)
}
\description{
Apply the Granger Causality test on both directions
}
