% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_of_group.R
\name{average_of_group}
\alias{average_of_group}
\title{Create a timeseries mean based on all the samples in the data}
\usage{
average_of_group(df = NULL)
}
\arguments{
\item{df}{a data.frame where the first column is a POSIXct object and every
subsequent column is a sample.}
}
\value{
returns the inputted data.frame with a mean column added.
}
\description{
Create a timeseries mean based on all the samples in the data
}
\examples{
\dontrun{
average_of_group(df = monitor_downsampled)
}

}
