% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_outputs.R
\name{bind_processed}
\alias{bind_processed}
\title{Bind Rows of Processed data}
\usage{
bind_processed(df = NULL, export = FALSE, path = getwd())
}
\arguments{
\item{df}{the output from process_data}

\item{export}{If TRUE, will save the data.frame as a .csv in the current directory. If FALSE (default),
will return the bound data.frame.}

\item{path}{Where to save the file if export = TRUE. Defaults to current directory.}
}
\value{
A single data.frame with all individuals row-bound and an additional \code{ID} column, or \code{NULL} (invisibly) when \code{export = TRUE}.
}
\description{
Allows for easy export and binding of rows from the output of process_data.
}
\examples{
bind_processed(df = monitor_processed)

}
