% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadiandynamics-package.R
\docType{package}
\name{circadiandynamics-package}
\alias{circadiandynamics}
\alias{circadiandynamics-package}
\title{circadiandynamics: Period and Rhythm Analysis of Timeseries Data}
\description{
Finds the period of univariate and multivariate timeseries (datetime) data by window. It will iterate over each window and find any changes in the period. Also fits a COSINOR function to the data.
}
\author{
\strong{Maintainer}: Eddie Pérez Claudio \email{eddieperezclaudio@gmail.com}

Authors:
\itemize{
  \item Eddie Pérez Claudio \email{eddieperezclaudio@gmail.com}
  \item Manuel Guiannoni-Guzmán
  \item Jonathan Alemán
  \item José L. Agosto-Rivera
}

}
\keyword{internal}
