% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_data.R
\name{crop_data}
\alias{crop_data}
\title{Crop data to specific dates}
\usage{
crop_data(df = NULL, from = NULL, to = NULL)
}
\arguments{
\item{df}{a list of data.frame/tibble with a column named datetime which is also a datetime object.}

\item{from, to}{a string which indicates a date in the ISO 8601 format.}
}
\value{
A list of data.frames/tibbles with dates within the specified window of time.
}
\description{
Crop data to specific dates
}
\details{
A helper function wrapped around dplyr::filter and lubridate::parse_date_time which helps to data to
datetimes of interest.
}
