% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_event.R
\name{define_event}
\alias{define_event}
\title{Define an Event for the USDA RFID data}
\usage{
define_event(df)
}
\arguments{
\item{df}{data.frame of RFID data containing at least the following columns: UID, UTCTime, Address, ScanCount.}
}
\value{
returns a data.frame with an additional column called "event".
}
\description{
This function takes data from a specific USDA project. It's not meant to be used with any other datasets.
}
\examples{
\dontrun{
define_event(df)
}

}
\author{
Eddie Perez Claudio
}
