% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_all.R
\name{export_all}
\alias{export_all}
\title{Export plots and data outputed by the package}
\usage{
export_all(processed_data = NULL, rythm_analysis_data = NULL, autocorrelation = TRUE, lomb_scargle = TRUE,
cosinor_fit = c("lomb_scargle", "autocorrelation"), dir_choose_gui = TRUE, new_dir_name = "analysis")
}
\arguments{
\item{processed_data}{data.frame returned from the process_timeseries or multivariate_process_timeseries functions.}

\item{rythm_analysis_data}{data.frame returned from the rythm_analysis_by_window and multivariate_rythm_analysis functions.}

\item{autocorrelation}{If TRUE (default) plots autocorrelation. FALSE does not plot autocorrelation.}

\item{lomb_scargle}{If TRUE (default) plots lomb scargle periodogram. FALSE does not plot periodogram.}

\item{cosinor_fit}{Method to fit the COSINOR to the data. Either "lomb_scargle" (default) or "autocorrelation".}

\item{dir_choose_gui}{If TRUE (default) a GUI will help select the folder in which to save the data and plots. If FALSE,
everything will be saved in the current directory.}

\item{new_dir_name}{Optional character argument to name the folder in which the data will be saved.}
}
\description{
Creates a folder in which all the data and plots resulting from the analysis will be saved.
}
\examples{
export_all(processed_data = processed_data, rythm_analysis_data = analysis_data, new_dir_name = "test")
}
