% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{Export data}
\usage{
export_data(processed_data = NULL, rythm_analysis_data = NULL, dir_choose_gui = TRUE)
}
\arguments{
\item{processed_data}{optional data.frame returned from the process_timeseries function.}

\item{rythm_analysis_data}{optional data.frame returned from the rythm_analysis_by_window function.}

\item{dir_choose_gui}{If TRUE (default) a GUI will help select the folder in which to save the data and plots. If FALSE,
everything will be saved in the current directory.}
}
\description{
Saves data outputted from the process_timeseries and rythm_analysis_by_window functions.
}
\examples{
export_data(processed_data = processed_data, rythm_analysis_data = analysis_data, dir_choose_gui = TRUE)
}
