% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multivariate_data.R
\name{export_multivariate_data}
\alias{export_multivariate_data}
\title{Export multivariate data}
\usage{
export_multivariate_data(processed_data = NULL, rythm_analysis_data = NULL, dir_choose_gui = TRUE, new_dir_name = "analysis")
}
\arguments{
\item{processed_data}{optional data.frame returned from the process_timeseries function.}

\item{rythm_analysis_data}{optional data.frame returned from the rythm_analysis_by_window function.}

\item{dir_choose_gui}{If TRUE (default) a GUI will help select the folder in which to save the data and plots. If FALSE,
everything will be saved in the current directory.}

\item{new_dir_name}{Optional character argument to name the folder in which the data will be saved.}
}
\description{
Saves data outputted from the multivariate process_timeseries and multivariate_rythm_analysis functions.
}
\examples{
export_multivariate_data(processed_data = processed_data, rythm_analysis_data = analysis_data,
dir_choose_gui = TRUE, new_dir_name = "Sensor_project")

}
