% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figures.R
\name{export_plots}
\alias{export_plots}
\title{Draw and Export plots from the analysis}
\usage{
export_plots(filename = "processed_and_analyzed_data.pdf", processed_data = NULL,
rythm_analysis_data = NULL, autocorrelation = TRUE,
lomb_scargle = TRUE, cosinor_fit = c("lomb_scargle", "autocorrelation"),
dir_choose_gui = TRUE)
}
\arguments{
\item{filename}{A charater string to name the ".pdf" file outputted. Default = "processed_and_analyzed_data".}

\item{processed_data}{data.frame returned from the process_timeseries or multivariate_process_timeseries functions.}

\item{rythm_analysis_data}{data.frame returned from the rythm_analysis_by_window and multivariate_rythm_analysis functions.}

\item{autocorrelation}{If TRUE (default) plots autocorrelation. FALSE does not plot autocorrelation.}

\item{lomb_scargle}{If TRUE (default) plots lomb scargle periodogram. FALSE does not plot periodogram.}

\item{cosinor_fit}{Method to fit the COSINOR to the data. Either "lomb_scargle" (default) or "autocorrelation".}

\item{dir_choose_gui}{If TRUE (default) a GUI will help select the folder in which to save the data and plots. If FALSE,
everything will be saved in the current directory.}
}
\description{
Draw and Export plots from the analysis
}
\examples{
export_plots(processed_data = df_processed, rythm_analysis_data = df_analysis)

}
