% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gaps.R
\name{find_gaps}
\alias{find_gaps}
\title{Fill Gaps in the data}
\usage{
find_gaps(times = NULL, sampling_rate = NULL)
}
\arguments{
\item{times}{a POSIXct vector.}

\item{sampling_rate}{A character string indicating the sampling rate of the data. Examples: '30 minutes', '1 hour', '4 seconds', '100 days'.}
}
\value{
POSIXct vector without gaps in the dates.
}
\description{
Finds and fills gaps in a POSIXct vector.
}
\examples{
\dontrun{
completed_dates <- find_gaps(times = df$datetime, sampling_rate = "30 min")
}

}
