% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_satelite.R
\name{import_satellite}
\alias{import_satellite}
\title{Import Weather data From satellite}
\usage{
import_satellite(path = NULL)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Import Weather data From satellite
}
\examples{

}
