% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_sampling_rate.R
\name{infer_sampling_rate}
\alias{infer_sampling_rate}
\title{Infer sampling rate}
\usage{
infer_sampling_rate(dates = NULL)
}
\arguments{
\item{dates}{POSIXct vector.}
}
\value{
int or dlb for the sampling mode of the POSIXct vector.
}
\description{
Attempts to approximate sampling rate by selecting the sampling mode.
}
\examples{
inferred_rate <- infer_sampling_rate(dates = df$datetime)
}
