% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time_windows.R
\name{make_time_windows}
\alias{make_time_windows}
\title{Create iterable time windows}
\usage{
make_time_windows(data = NULL, window_size_in_days = 3, window_step_in_days = 1)
}
\arguments{
\item{data}{a data.frame with 2 columns. Column 1 must be a POSIXct object and column 2 must be the measurement values.}

\item{window_size_in_days}{a numeric indicating the width of the window size in day units.}

\item{window_step_in_days}{a numeric indicating the amount of day by which to move the window in day units.}
}
\value{
A data.frame cotaining:

window     iterator by which to subset the data.frame
values     raw measurement data
}
\description{
Creates iterable time windows for a data.frame
}
\examples{
\dontrun{
windowed_data <- make_time_windows(data = df)
}

}
