% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actogram_temperature.R
\name{plot_actogram_temp}
\alias{plot_actogram_temp}
\title{Export/Plot Actogram with Temperature}
\usage{
plot_actogram(df = NULL, ld_data = NULL, datetime_column = 1, filename = "actogram.pdf",
export = FALSE, width = 12, height = 12, dpi = 800, nrow = 5, ncol = 5)
}
\arguments{
\item{df}{A data.frame or tibble which contains a datetime column and measurement values.}

\item{ld_data}{A data.frame/Tibble with 2 columns. Column 1 is a datetime object and column 2 is the light/dark indicator. (defult = NULL)}

\item{datetime_column}{An integer which indicates the column that contains the datetime. Default = 1.}

\item{filename}{A charater string  ending with ".pdf" which customizes the name of the file to be exported. "actogram.pdf" (default)}

\item{export}{Logical. If FALSE (default), it will open a window with the plots. If TRUE, saves the plots on the current directory in pdf format.}

\item{width}{a numeric indicating the width of the page of the pdf file. default = 12}

\item{height}{a numeric indicating height of the page of the pdf file. default = 12}

\item{nrow}{a numeric indicating how many rows of plots to put in a page. default = 5}

\item{ncol}{a numeric indicating how many columns of plots to put in a page. default = 5}

\item{autosize}{Logical. If TRUE, will make the figures bigger as the number of days increase. If FALSE
(default) will plot 25 figures per page.}
}
\description{
Export/Plot Actogram with Temperature
}
\details{
Exports actogram plots for each column of a data.frame of tibble with a datetime object.
}
\examples{
plot_actogram(df = monitor,
datetime_column = 1, filename = "actogram.pdf",
export = FALSE)

}
