% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actogram_windows.R
\name{plot_actogram_windows}
\alias{plot_actogram_windows}
\title{Plot Window Actogram}
\usage{
plot_actogram_windows(df)
}
\arguments{
\item{df}{The 'data' output from simplify_data}
}
\value{
a list of actograms binned by window
}
\description{
Plot Window Actogram
}
