% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_autocorrelation.R
\name{plot_autocorrelation}
\alias{plot_autocorrelation}
\title{Plot the Autocorrelation by windows}
\usage{
plot_autocorrelation(df, path = NULL, filename = "actogram.pdf", dir_choose_gui = TRUE,
export = TRUE, window_of_interest = NULL)
}
\arguments{
\item{df}{a data.frame containing the output from acf_window.}

\item{path}{a string containing a path in which to save the file. It is not necesssary if dir_choose_gui = TRUE.}

\item{filename}{a string containing a name for the file. It must end in ".pdf". Default = "actogram.pdf".}

\item{dir_choose_gui}{A gui for choosing the directory/folder in which to save the file. If, false, a path
is needed. Default = TRUE.}

\item{export}{If FALSE, will plot the figures in the console. If TRUE, will export the figures as a .pdf file.}

\item{window_of_interest}{optional. a number or list of numbers which indicate the windows to be plotted.}
}
\value{
Invisibly returns a list of \code{ggplot} objects, or saves a PDF when \code{export = TRUE}.
}
\description{
Plot the Autocorrelation by windows
}
