% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_non_ld_actogram.R
\name{plot_non_ld_actogram}
\alias{plot_non_ld_actogram}
\title{Internal Function to plot Non LD Actograms}
\usage{
plot_non_ld_actogram(df, id)
}
\arguments{
\item{df}{The output of internal formatting in the plot_actogram2 function.}
}
\value{
A set of ggplot2 actograms
}
\description{
An internal helper function not meant to be called by the user.
}
