% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase.R
\name{plot_phase}
\alias{plot_phase}
\title{Phase plots for all individuals}
\usage{
plot_phase(path = getwd(), analysis = NULL)
}
\arguments{
\item{path}{directory for where to save the plots}

\item{analysis}{output from the analysis made with Circadian Dynamics}
}
\value{
a pdf with plots
}
\description{
Phase plots for all individuals
}
\examples{
\dontrun{
plot_phase(path = getwd(), analysis = analysis_data)
}

}
