% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_window_data.R
\name{plot_window_data}
\alias{plot_window_data}
\title{Utility Function for Plotting Detaiiled data in Windows}
\usage{
plot_window_data(df, alt_cos = FALSE)
}
\arguments{
\item{df}{the output from 'simplify_data'.}

\item{alt_cos}{Logical. If TRUE, uses the alternative cosinor fit when plotting. Default = FALSE.}
}
\value{
a list of plots.
}
\description{
Utility Function for Plotting Detaiiled data in Windows
}
\examples{
\dontrun{
window_plots = plot_window_data(df)
}
}
