% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_window_data_app.R
\name{plot_window_data_app}
\alias{plot_window_data_app}
\title{Utility Function for Plotting Detaiiled data in Windows for the App}
\usage{
plot_window_data_app(df)
}
\arguments{
\item{df}{the output from 'simplify_data'.}
}
\value{
a list of plots.
}
\description{
Utility Function for Plotting Detaiiled data in Windows for the App
}
\examples{
\dontrun{
window_plots = plot_window_data_app(df)
}
}
