% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clocklab.R
\name{read_clocklab}
\alias{read_clocklab}
\alias{read_clocklab_folder}
\alias{read_clocklab_interactive}
\alias{read_clocklab_folder_interactive}
\title{Import data from Clocklab}
\usage{
read_clocklab(file = NULL)

read_clocklab_folder(directory = NULL)

read_clocklab_interactive(...)

read_clocklab_folder_interactive(...)
}
\arguments{
\item{file}{A file path for a .csv outputted from the clocklab software.}

\item{directory}{A folder path containing .csv files outputted from the clocklab software.
Make sure all individuals in the folder belong to the same experimental group and that the experiments
were run on the same dates.}
}
\value{
Returns a data.frame/tibble with 3 columns:
datatetime of the experiment
ld : light switch status
IND Name: measurement values
}
\description{
Import data from Clocklab
}
\examples{
df <- read_clocklab(file = "/path/to/file.csv")
df <- read_clocklab_folder(directory = "/path/to/folder")

}
