% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale the data from 0 to 1.}
\usage{
rescale(x)
}
\arguments{
\item{x}{A numeric vector to rescale.}
}
\value{
A numeric vector rescaled to the range \code{[0, 1]}.
}
\description{
Rescale the data from 0 to 1.
}
