% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_inactive.R
\name{rm_inactive}
\alias{rm_inactive}
\title{Remove Inactive Individuals}
\usage{
rm_inactive(df, inactivity_period = "1 day", sampling_rate = "1 hour")
}
\arguments{
\item{df}{a data.frame where the first column is a datetime object.}

\item{inactivity_period}{a string indicating what will be considered the period of inactivity
necessary to be removed. Default = "1 day".}

\item{sampling_rate}{a string indicating the sampling rate of the data. Default = "1 hour".}
}
\value{
A data.frame with only the active individuals (columns) retained.
}
\description{
Remove Inactive Individuals
}
