% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rythm_analysis_by_window_2.R
\name{rythm_analysis_by_window_2}
\alias{rythm_analysis_by_window_2}
\title{Helper function to analyze the period of a time series}
\usage{
multivariate_rythm_analysis(df = NULL, sampling_rate = NULL, auto_correlation = TRUE, lomb_scargle = TRUE,
from = NULL, to = NULL, ofac = 1, multipeak_period = TRUE, peak_of_interest = Inf,
datetime = NULL, window = NULL, values = NULL)
}
\arguments{
\item{df}{The output from multivariate_process_timeseries}

\item{sampling_rate}{A character string indicating the sampling rate of the data. Examples: '30 minutes', '1 hour', '4 seconds', '100 days'.}

\item{from}{An optional numeric indicating from which period or frequency to start looking for peaks.
Must be in hours. Default = 18.}

\item{to}{An optional numeric indicating up to which period or frequency to start looking for peaks.
Must be in hours. Default = 30.}

\item{ofac}{[lomb::lsp()] The oversampling factor. Must be an integer>=1. Larger values of ofac lead to finer scanning of frequencies but may be time-consuming for large datasets and/or large frequency ranges (from...to).}

\item{autocorrelation}{Logical. If TRUE (default) runs an autocorrelation on each window of the data. If FALSE, does not run.}

\item{lomb_scargle}{Logical. If TRUE (default) runs the Lomb-Scargle Periodogram on each window on the data. If FALSE, does not run.}

\item{datetime}{optional if a data.frame is supplied. A list of multiple POSIXct vectors corresponding to each mearument variable.}

\item{window}{optional if a data.frame is supplied. A list of multiple window vectors corresponding to each mearument variable.}

\item{values}{optional if a data.frame is supplied. A list of multiple value vectors corresponding to each mearument variable.}
}
\value{
A names list of data.frames with the combined output of [acf_by_window()], [lsp_by_window()], and [cosinor_lm()].
}
\description{
Helper function which takes the output from process_timeseries and runs: autocorrelation,
lomb-scargle periodogram, and COSINOR analysis.
}
\examples{
analysis <- rythm_analysis(df = processed_data, sampling_rate = "30 min", auto_correlation = TRUE,
lomb_scargle = TRUE, from = 36, to = 48, multipeak_period = TRUE)

}
