% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_data.R
\name{truncate}
\alias{truncate}
\title{Truncate data}
\usage{
truncate(df = NULL, from = NULL, to = NULL)
}
\arguments{
\item{df}{a data.frame/tibble with a column named datetime which is also a datetime object.}

\item{to}{a string which indicates a date in the ISO 8601 format.}
}
\value{

}
\description{
Truncate data
}
\details{
A helper function wrapped around dplyr::filter and lubridate::parse_date_time which helps to truncate datetime
data.
}
