% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actogram.R
\name{actogram}
\alias{actogram}
\title{Actogram}
\usage{
actogram(df, sampling, file = "actogram")
}
\arguments{
\item{df}{First column is datetime, the rest are inds.}

\item{file}{}
}
\description{
Actogram
}
