% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_timeseries.acf.R
\name{analyze_acf}
\alias{analyze_acf}
\title{Auto Correlation with a moving window}
\usage{
analyze_acf(df = NULL, from = 18, to = 30, sampling_rate = "1 hour")
}
\arguments{
\item{df}{A data.frame with 2 columns. Column 1 must contain the windows to iterate over.
Column 2 must supply the values. This parameter is optional if window_vector and values are supplied.
df must not have gaps in the dates, acf asumes data is evenly spaced.}

\item{from}{The period (in hours) from which to start looking for peaks in the autocorrelation. Default = 18.}

\item{to}{The period (in hours) up to which  to look for peaks in the autocorrelation. Default = 30.}

\item{sampling_rate}{A charater string which indicates the sampling rate of the data.
For example: "1 second", "2 minutes", "1 hour" (default),"3 days", "11 months".}
}
\value{
A data.frame with the autocorrelation results for each window which include: period, peaks,
power, lags for the peaks.
}
\description{
Uses autocorrelation to find a circadian period for a given timeseries
}
\details{
The reported `rhythm_strength` is the maximum autocorrelation peak within the
search band divided by the 95% white-noise confidence bound \code{1.965 / sqrt(n)}, where \code{n} is the number of observations in \code{df}. Values > 1
indicate the peak exceeds what white noise would produce. The Lomb-Scargle path
reports a different quantity, `relative_power` (in [analyze_lomb]), which is on a
different scale and is not directly comparable.
}
\examples{
\dontrun{
res <- analyze_acf(df, from = 18, to = 30, sampling_rate = "1 hour")
}

}
\seealso{
[stats::acf()] which this functions uses to run the autocorrelation.
}
