% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_timeseries.cosinor.R
\name{analyze_cosinor}
\alias{analyze_cosinor}
\title{Cosinor Analysis}
\usage{
analyze_cosinor(
  df = NULL,
  sampling_rate = NULL,
  period = NULL,
  na.action = na.omit
)
}
\arguments{
\item{df}{A data.frame which contains the dates of a time series in column 1 and the values in column 2.}

\item{sampling_rate}{A character string indicating the sampling rate of the data. Examples: '30 minutes', '1 hour', '4 seconds', '100 days'.}

\item{period}{A numeric indicating the period to analyse. Must be in the same units as the sampling rate.
Examples: If the goal is to evaluate a 24 "hour" period but the sampling rate is "30 minutes",
the period to use is a  48 "30 minutes" period.}

\item{na.action}{Default is na.omit which excludes NA values from the analysis. See [stats::lm()] for a more detailed description.}
}
\value{
A data.frame with:
MESOR: The intercept of the regression. ie. Mean of the COSINOR fit.
Amplitude: Amplitude of the COSINOR fit.
Amplitude_se
Acrophase: Phase shifting of the fit in radians.
Acrophase_se
Phase_in_seconds: Acrophase in seconds.
Phase_se_seconds
Adj_r_squared: The PR or Percent Rythm of the fit.
Cosinor_p_value: p.value for the regression. Significance indicates the period matches the data.
Wave_y = MESOR + amplitude * cosw : The use is to plot the fit.
Wave_x = timeseries_datetime + phase_in_seconds : The use is to plot the fit.
}
\description{
COSINOR analysis of a timeseries.
}
\examples{
\dontrun{
res <- analyze_cosinor(df, sampling_rate = "1 hour", period = 24)
}

}
\seealso{
Barnett, A. G., & Dobson, A. J. (2010).
Analysing Seasonal Health Data. Statistics for Biology and Health.
doi:10.1007/978-3-642-10748-1

Tong, Y. L. (1976).
Parameter Estimation in Studying Circadian Rhythms.
Biometrics, 32(1), 85. doi:10.2307/2529340
}
