% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbitrary_noise_substraction.R
\name{arbitrary_noise_subtraction}
\alias{arbitrary_noise_subtraction}
\title{Subtract Arbitrarily Chosen Noise Level from the Data}
\usage{
arbitrary_noise_reduction(df, noise = 0)
}
\arguments{
\item{df}{a data.frame where the first column is a datetime object.}

\item{noise}{an integer representing an arbitrarily chosen noise level to subtract from the data.
Default = 0.}
}
\value{

}
\description{
Subtract Arbitrarily Chosen Noise Level from the Data
}
