% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_outputs.R
\name{bind_processed}
\alias{bind_processed}
\title{Bind Rows of Processed and Analyzed data}
\usage{
bind_processed(df = NULL, export = FALSE, path = paste0(getwd(),"/",substitute(df),".csv"))
bind_analysis(df = NULL, export = FALSE, path = paste0(getwd(),"/",substitute(df),".csv"))
}
\arguments{
\item{df}{the output from either process_data or rythm_analysis}

\item{export}{If TRUE, will save the data.frame as a .csv in the current directory. If FALSE (default),
will return the bound data.frame.}

\item{path}{Where to save the file if export = TRUE. Defaults to current directory.}
}
\description{
Allows for easy export and binding of rows from the outputs of process_data and rythm_analysis.
}
\examples{
bind_processed(df = monitor_processed)
bind_analysis(df = monitor_analysis)
}
