% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterworth_filter.R
\name{butterworth_filter}
\alias{butterworth_filter}
\title{Butterworth Filter Helper}
\usage{
butterworth_filter(
  df = NULL,
  order = 2,
  f_low = 1/4,
  f_high = 1/72,
  plot = TRUE,
  ...
)
}
\arguments{
\item{df}{required. A data.frame object where column 1 is a POSIXct object and the other columns are measurement values.}

\item{order}{filter order. Default = 2.}

\item{f_low}{Frequency for the low pass filter. Default = 1/4.}

\item{f_high}{Frequency for the high pass filter. Default = 1/72.}

\item{plot}{logical. If TRUE (default) plots the filtered data over the raw data. Red line is the low pass filter. Blue is the high pass filter. If FALSE, does not plot.}
}
\value{
The input data.frame \code{df} with an additional column \code{butterworth} containing the filtered values.
}
\description{
A helper function which wraps around the [signal::butter()] and [signal::filfilt()] functions.
It can be either a low or high pass filter for a given period. (The function can take the period argument
because it does the conversion to frequency automatically. (frequency = 1/period).
}
\examples{

butter <- butterworth_filter(df = data, f_low = 1/4, f_high = 1/72)
print(butter)

}
