% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterworth_filter.R
\name{butterworth_filter}
\alias{butterworth_filter}
\title{Butterworth Filter Helper}
\usage{
butterworth_filter(df = NULL, period = 24,
     type = c("low", "high"), plot = TRUE)
}
\arguments{
\item{df}{required. A data.frame object where column 1 is a POSIXct object and the other columns are measurement values.}

\item{period}{The period on which we want to draw the threshold (default = 24).}

\item{type}{A string indicating if a low or high pass filter will be used. options: "low" (default), "high".}

\item{plot}{logical. If TRUE (default) plots the filtered data over the raw data. If FALSE, does not plot.}
}
\description{
A helper function which wraps around the [signal::butter] and [signal::filfilt] functions.
It can be either a low or high pass filter for a given period. (The function can take the period argument
because it does the conversion to frequency automatically. (frequency = 1/period).
}
\examples{

butter <- butterworth_filter(df = data,
           period = 24, type = "low", plot = FALSE)
print(butter)

}
