% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{circadiandynamics-deprecated}
\alias{circadiandynamics-deprecated}
\alias{read_trikinetics_nested}
\alias{read_trikinetics_long}
\alias{read_trikinetics_folder_nested}
\alias{read_trikinetics_folder_long}
\alias{read_trikinetics_nested_interactive}
\alias{read_trikinetics_long_interactive}
\alias{read_trikinetics_folder_nested_interactive}
\alias{read_trikinetics_folder_long_interactive}
\alias{read_clocklab_folder}
\alias{read_clocklab_interactive}
\alias{read_clocklab_folder_interactive}
\alias{read_csv_data_interactive}
\alias{read_satellite_interactive}
\alias{read_vitalpatch_interactive}
\alias{analyze_timeseries.acf}
\alias{analyze_timeseries.cosinor}
\alias{analyze_timeseries.lomb}
\alias{process_timeseries.main}
\alias{process_timeseries.core}
\alias{process_timeseries.waveform}
\title{Deprecated reader aliases}
\usage{
read_trikinetics_nested(file = NULL)

read_trikinetics_long(file = NULL)

read_trikinetics_folder_nested(directory = NULL)

read_trikinetics_folder_long(directory = NULL)

read_trikinetics_nested_interactive()

read_trikinetics_long_interactive()

read_trikinetics_folder_nested_interactive()

read_trikinetics_folder_long_interactive()

read_clocklab_folder(directory = NULL)

read_clocklab_interactive()

read_clocklab_folder_interactive()

read_csv_data_interactive()

read_satellite_interactive()

read_vitalpatch_interactive()

analyze_timeseries.acf(...)

analyze_timeseries.cosinor(...)

analyze_timeseries.lomb(...)

process_timeseries.main(...)

process_timeseries.core(...)

process_timeseries.waveform(...)
}
\arguments{
\item{path, file, directory, folder}{Path forwarded to the replacement function.}
}
\value{
The value of the replacement function.
}
\description{
These wrappers are kept for backward compatibility and will be
  removed in a future release. The reader API is now a single function per
  format that opens a dialog when its path is `NULL` and detects folders
  automatically: [read_trikinetics()], [read_clocklab()], [read_csv_data()],
  [read_satellite()], [read_vitalpatch()].
}
\keyword{internal}
