% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_event.R
\name{define_event}
\alias{define_event}
\title{Define an Event for the USDA RFID data}
\usage{
define_event(df, core = TRUE)
}
\arguments{
\item{df}{data.frame of RFID data containing at least the following columns: UID, UTCTime, Address, ScanCount.}

\item{core}{boolean. If TRUE (default), will run paralelization using the multicore approach. This doesn't work well on Windows 10
systems. If FALSE, it will use the multisession approach.}
}
\value{
returns a data.frame with an additional column called "event".
}
\description{
This function takes data from a specific USDA project. It's not meant to be used with any other datasets.
}
\examples{
define_event(df)

}
\author{
Eddie Perez Claudio
}
