% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_plots.R
\name{detailed_plots}
\alias{detailed_plots}
\title{Generate Detailed Plots (Utility function)}
\usage{
detailed_plots(trikinetics_analyzed, sampling_rate = "minutes", windows = TRUE)
}
\arguments{
\item{windows}{}
}
\description{
Generate Detailed Plots (Utility function)
}
