% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_time_series.R
\name{downsample_time_series}
\alias{downsample_time_series}
\title{Downsample a timeseries with a datetime object}
\usage{
downsample_time_series(
  data = NULL,
  amount = 1,
  units = "hour",
  method = c("mean", "sum", "median")
)
}
\arguments{
\item{data}{a list of data frames or tibbles, each with a POSIXct datetime column.}

\item{amount}{the amount of units to downsample for}

\item{units}{a character object like:"hour", "minute", "day", "week". Default is "hour"}

\item{method}{methods by which to summarise the measurement values in the timeseries: "mean", "sum", "median". Default is "mean".}
}
\value{
A list of data frames with downsampled data.
}
\description{
This function will resample time series data (downsample) to the nearest minute, half-hour, hour, or day.
It allows to summarise the measurement values using mean, median, or sum.
}
\examples{
\dontrun{
df_downsampled <- downsample_time_series(data = raw_data,
amount = 30, units = "hour", method = "sum")
}

}
