% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_time_series.R
\name{downsample_time_series}
\alias{downsample_time_series}
\title{Downsample a timeseries with a datetime object}
\usage{
downsample_time_series(data = NULL, datetime_column = "datetime",amount = 30,
       units = c("minute", "hour", "day", "week"), method = c("mean", "sum", "median"))
}
\arguments{
\item{data}{data frame or tibble with 2 columns. One column must be a POSIXct object.}

\item{datetime_column}{column in data that has the datetime}

\item{amount}{the amount of units to downsample for}

\item{units}{a character object like:"hour", "minute", "day", "week". Default is "minute"}

\item{method}{methods by which to summarise the measurement values in the timeseries: "mean", "sum", "median". Default is "mean".}
}
\value{
A data.frame with 2 columns:
downsampled_data: summarise measurement values
floored_dates: downsampled POSIXct object
}
\description{
This function will resample time series data ( downsample) to the nearest minute, half-hour, hour, or day.
It allow to summarise the measurement values using mean, median, or sum.
}
\examples{
df_downsampled <- downsample_time_series(data = raw_data,
datetime_column = "datetime", amount = 30, units = "hour", method = "sum")

}
