% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots_with_windows.R
\name{generate_plots_with_windows}
\alias{generate_plots_with_windows}
\title{Function to Generate Plots of the Analyzed Data
Must be used when 'Make Time Windows = TRUE'}
\usage{
generate_plots_with_windows(trikinetics_analyzed, sampling_rate = "minutes")
}
\description{
Function to Generate Plots of the Analyzed Data
Must be used when 'Make Time Windows = TRUE'
}
