% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_peaks.R
\name{lsp_peaks}
\alias{lsp_peaks}
\title{Find peaks in an lsp object}
\usage{
lsp_peaks(object, npeaks = 5)
}
\arguments{
\item{object}{An lsp object returned by \code{lsp_mod()}.}

\item{npeaks}{Integer. Maximum number of peaks to return. Default is 5.}
}
\value{
A data.frame with columns \code{time} (period/frequency at each peak) and \code{peaks} (peak power).
}
\description{
Find peaks in an lsp object
}
\keyword{internal}
