% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overplot_actogram.R
\name{overplot_actogram}
\alias{overplot_actogram}
\title{Export/Plot Actogram with Temperature}
\usage{
overplot_actogram(
  path = getwd(),
  df = NULL,
  overplot = NULL,
  ld_data = NULL,
  datetime_column = 1,
  filename = "actogram.pdf",
  export = FALSE,
  width = 12,
  height = 12,
  dpi = 800,
  nrow = 5,
  ncol = 5,
  autosize = FALSE
)
}
\arguments{
\item{df}{A data.frame or tibble which contains a datetime column and measurement values.}

\item{overplot}{A list or vector containing the values to be overplotted on the actograms. It must be of the same dimensions/binning as the df.}

\item{ld_data}{A data.frame/Tibble with 2 columns. Column 1 is a datetime object and column 2 is the light/dark indicator. (defult = NULL)}

\item{datetime_column}{An integer which indicates the column that contains the datetime. Default = 1.}

\item{filename}{A charater string  ending with ".pdf" which customizes the name of the file to be exported. "actogram.pdf" (default)}

\item{export}{Logical. If FALSE (default), it will open a window with the plots. If TRUE, saves the plots on the current directory in pdf format.}

\item{width}{a numeric indicating the width of the page of the pdf file. default = 12}

\item{height}{a numeric indicating height of the page of the pdf file. default = 12}

\item{nrow}{a numeric indicating how many rows of plots to put in a page. default = 5}

\item{ncol}{a numeric indicating how many columns of plots to put in a page. default = 5}

\item{autosize}{Logical. If TRUE, will make the figures bigger as the number of days increase. If FALSE
(default) will plot 25 figures per page.}
}
\value{
Invisibly returns a multi-page \code{grob} object (from \code{gridExtra::marrangeGrob}), or saves a PDF when \code{export = TRUE}.
}
\description{
Export/Plot Actogram with Temperature
}
\details{
Exports actogram plots for each column of a data.frame of tibble with a datetime object.
}
\examples{
overplot_actogram(df = monitor, overplot = monitor[[3]],
datetime_column = 1, filename = "actogram.pdf",
export = FALSE)

}
