% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf_results.R
\name{plot_acf_results}
\alias{plot_acf_results}
\title{Plot Raw Values}
\usage{
plot_acf_results(df)
}
\arguments{
\item{df}{Use the second output from 'simplify_data'.}
}
\value{
Returns a list of plots for the autocorrelation results.
}
\description{
Plot Raw Values
}
\examples{
plots = plot_acf_results(df = simplified_data$autocorrelation)

}
