% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actogram.R
\name{plot_actogram}
\alias{plot_actogram}
\title{Export/Plot Actogram}
\usage{
plot_actogram(df = NULL, datetime_column = 1, filename = "actogram.pdf")
}
\arguments{
\item{df}{A data.frame or tibble which contains a datetime column and measurement values.}

\item{datetime_column}{An integer which indicates the column that contains the datetime.}

\item{filename}{A charater string  ending with ".pdf" which customizes the name of the file to be exported. "actogram.pdf" (default)}
}
\description{
Export/Plot Actogram
}
\details{
Exports actogram plots for each column of a data.frame of tibble with a datetime object.
}
\examples{
plot_actogram(df = monitor,
datetime_column = 1)
}
