% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actogram.R
\name{plot_actogram}
\alias{plot_actogram}
\title{Export/Plot Actogram}
\usage{
plot_actogram(df = NULL, datetime_column = 1, filename = "actogram.pdf",
export = FALSE, width = 12, height = 12, dpi = 800, nrow = 5, ncol = 5)
}
\arguments{
\item{df}{A data.frame or tibble which contains a datetime column and measurement values.}

\item{datetime_column}{An integer which indicates the column that contains the datetime.}

\item{ld_column}{An integer which indicates the column that contains the light/dark data. (defult = NULL)}

\item{filename}{A charater string  ending with ".pdf" which customizes the name of the file to be exported. "actogram.pdf" (default)}

\item{export}{Logical. If FALSE (default), it will open a window with the plots. If TRUE, saves the plots on the current directory in pdf format.}

\item{width}{a numeric indicating the width of the page of the pdf file. default = 12}

\item{height}{a numeric indicating height of the page of the pdf file. default = 12}

\item{nrow}{a numeric indicating how many rows of plots to put in a page. default = 5}

\item{ncol}{a numeric indicating how many columns of plots to put in a page. default = 5}
}
\description{
Export/Plot Actogram
}
\details{
Exports actogram plots for each column of a data.frame of tibble with a datetime object.
}
\examples{
plot_actogram(df = monitor,
datetime_column = 1, filename = "actogram.pdf",
export = FALSE)

}
