% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actogram2.R
\name{plot_actogram2}
\alias{plot_actogram2}
\title{Plot Actograms}
\usage{
plot_actogram2(df)
}
\arguments{
\item{df}{The 'data' output from simplify_data}
}
\value{
a list of actograms
}
\description{
Plot Actograms
}
