% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lsp_results.R
\name{plot_lsp_results}
\alias{plot_lsp_results}
\title{Plot Lomb-Scargle Results}
\usage{
plot_lsp_results(df)
}
\arguments{
\item{df}{Use the third output from 'simplify_data'.}
}
\value{
Returns a list of plots for the autocorrelation results.
}
\description{
Plot Lomb-Scargle Results
}
\examples{
plots = plot_lsp_results(df = simplified_data$lombscargle)

}
