% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_values.R
\name{plot_raw_values}
\alias{plot_raw_values}
\title{Plot Raw Values}
\usage{
plot_raw_values(df)
}
\arguments{
\item{df}{Use the first output from 'simplify_data'.}
}
\value{
Returns a list of plots for the raw values.
}
\description{
Plot Raw Values
}
\examples{
plots = plot_raw_values(df = simplified_data$data)

}
