% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_data.R
\name{read_csv_data}
\alias{read_csv_data}
\title{Import data from a Trikinetics tsv}
\usage{
read_csv_data(file = NULL)
}
\arguments{
\item{file}{a path to a csv file. If no path is supplied, a GUI will open and help with file selection.}
}
\description{
Import data from a Trikinetics tsv
}
\examples{
circadian_data <- read_csv_data()
}
