% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trikinetics.R
\name{read_trikinetics}
\alias{read_trikinetics}
\title{Import data from a Trikinetics tsv}
\usage{
read_trikinetics()
}
\description{
Import data from a Trikinetics tsv
}
\examples{
trikinetics_data <- read_trikinetics()

}
