% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trikinetics_nested.R
\name{read_trikinetics_nested}
\alias{read_trikinetics_nested}
\alias{read_trikinetics_folder_nested}
\alias{read_trikinetics_nested_interactive}
\alias{read_trikinetics_folder_nested_interactive}
\title{Import data from a Trikinetics tsv}
\usage{
read_trikinetics_nested(file = NULL)

read_trikinetics_folder_nested(directory = NULL)

read_trikinetics_nested_interactive(...)

read_trikinetics_folder_nested_interactive(...)
}
\arguments{
\item{file}{a path to a Trikinetics file.}

\item{directory}{a path to a folder containing Trikinetics files.}

\item{...}{Arguments passed on to \code{read_trikinetics_nested} or \code{read_trikinetics_folder_nested}.}
}
\value{
A named list of tibbles, one per individual, each with columns \code{datetime}, \code{ld}, and \code{value}.
}
\description{
Import data from a Trikinetics tsv
}
\examples{
\dontrun{
trikinetics_data <- read_trikinetics_nested(file = "/path/to/file.txt")
}
}
