% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vitalpatch.R
\name{read_vitalpatch}
\alias{read_vitalpatch}
\alias{read_vitalpatch_interactive}
\title{Read a Folder of Vital Patch Data}
\usage{
read_vitalpatch(folder = NULL)

read_vitalpatch_interactive(...)
}
\arguments{
\item{folder}{A path (directory) to the folder we want to analyze.}

\item{...}{Arguments passed on to \code{read_vitalpatch}.}
}
\value{
Returns a tibble (data.frame) object with parsed dates.
}
\description{
This function imports a folder of Vital Patch data in the standard format. That is,
the format provided in Vital Patch's Eval Data. There should be a folder per channel. read_vitalpatch
concatenates all the data into a single tibble (data.frame) so that the user doesn't have to do it manually.
}
\details{
Parallelism follows the caller's \code{future::plan()}; set a plan (e.g. \code{future::plan(future::multisession)}) before calling to parallelize.
}
\examples{
\dontrun{
read_vitalpatch(folder = "/path/to/folder")
}

}
