% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managing_inactives.R
\name{rm_inactive_dates}
\alias{rm_inactive_dates}
\title{Managing Inactive Variables}
\usage{
rm_inactive_dates(df,inactivity_period = "1 day", sampling_rate = "1 hour")

rm_inactive_variables(df,inactivity_period = "1 day", sampling_rate = "1 hour")

report_inactive_variables(df,inactivity_period = "1 day", sampling_rate = "1 hour")
}
\arguments{
\item{df}{a data.frame where the first column is a datetime object.}

\item{inactivity_period}{a string indicating what will be considered the period of inactivity
necessary to be removed. Default = "1 day".}

\item{sampling_rate}{a string indicating the sampling rate of the data. Default = "1 hour".}
}
\description{
rm_inactive_dates drops the rows of data during which an variable was "inactive" ie. equal to 0.

rm_inactive_variables drops the rows of data during which
}
\examples{
 df = downsample_time_series(trikinetics, amount = 1, units = 'hour', method = 'sum')
cropped_dates_df = rm_inactive_dates(df, inactivity_period = '1 day', sampling_rate = '1 hour')
print(cropped_dates_df)
cropped_variables_df = rm_inactive_variables(df,inactivity_period = "1 day", sampling_rate = "1 hour")
print(cropped_variables_df)

}
