% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plot_window_data.R
\name{shiny_plot_window_data}
\alias{shiny_plot_window_data}
\title{Utility Function for Plotting Detaiiled data by Window in the ShinyApp}
\usage{
shiny_plot_window_data(df)
}
\arguments{
\item{df}{the output from 'simplify_data'.}
}
\value{
a list of plots.
}
\description{
Utility Function for Plotting Detaiiled data by Window in the ShinyApp
}
\examples{
window_plots = plot_window_data(df)
}
