% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_data.R
\name{simplify_data}
\alias{simplify_data}
\title{Simplify Data}
\usage{
simplify_data(df, big_data = FALSE)
}
\arguments{
\item{df}{Output from processed_timeseries.main}

\item{big_data}{Are you using a large dataset? TRUE or FALSE (default)}
}
\description{
Simplify Data
}
