% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_and_detrend.R
\name{smooth_and_detrend}
\alias{smooth_and_detrend}
\title{Smooth and/or Detrend Data by Windows}
\usage{
smooth_and_detrend(
  df = NULL,
  smooth_data = TRUE,
  binning_n = 4,
  detrend_data = TRUE
)
}
\arguments{
\item{df}{a data.frame with 2 columns. The first column must be the windows. The second column the values to process.}

\item{smooth_data}{Logical. If TRUE (default) will smooth the measurement values useing a moving average. If FALSE measurement values won't be smoothed.}

\item{binning_n}{A numeric which indicated the amount of bins over which to run the smoothing average. Default = 4.}

\item{detrend_data}{Logical. If TRUE (default) will detrend the data. If FALSE measurement values won't be detrended. If both, detrend_data and smooth_data are TRUE, the detrending will run over the smoothed data.}

\item{windows}{Optional if a data.frame is supplied. A vector with the windows.}

\item{values}{Optional if a data.frame is supplied. A vector of values from a mesurement.}
}
\value{
A data.frame conatining: /n

window                 A vector with the windows. /n
values                 The raw measurement values. /n
smoothed               A column with the smoothed data. /n
detrended              A column with the detrended data. /n
smoothed_and_detrended A column with data that's been both smoothed and detrended.
}
\description{
Smooth and/or Detrend Data by Windows
}
\examples{
smoothed_data <- smooth_detrend_by_windows(df = windowed_data, smooth_data = TRUE, detrend_data = FALSE)

}
